Struct gooey_reactor::Dynamic

pub struct Dynamic<T>where
    T: 'static,{ /* private fields */ }

A handle to a dynamic value that can be updated and reacted to.

The T value is stored within the Scope that created the Dynamic. It can be updated using one of these approaches:

Dynamic::map_mut()

Dynamic::set()

Changes to this type can be observed using one of these approaches:

Dynamic::iter()

Dynamic::for_each()

Dynamic::map_each()

Dynamic::into_stream()

Implementations

impl<T> Dynamic<T>where T: 'static,

pub fn get(&self) -> Option<T>where T: Clone,

Returns a clone of the currently contained value.

None is returned if this value belonged to a scope that has been destructed.

pub fn map_ref<R>(&self, map: impl FnOnce(&T) -> R) -> Option<R>

Calls map with the currently contained value.

All other access to this value will be blocked while map executes.

None is returned if this value belonged to a scope that has been destructed.

pub fn map_mut<R>(&self, map: impl FnOnce(&mut T) -> R) -> Option<R>

Calls map with the currently contained value.

All other access to this value will be blocked while map executes.

None is returned if this value belonged to a scope that has been destructed.

pub fn set(&self, new_value: T) -> Option<T>

Replaces the currently contained value with new_value, returning the existing value.

None is returned if this value belonged to a scope that has been destructed.

pub fn iter(&self) -> ValueIterator<T>

Returns an iterator to values contained in this dynamic.

pub fn into_stream(self) -> ValueStream<T>

Returns an async [futures_core::Stream] returning values contained in this dynamic.

pub fn map_each<R, F>(&self, map: F) -> Option<Dynamic<R>>where F: for<'a> FnMut(&'a T) -> R + Send + 'static, R: Send + Sync,

Returns a new Dynamic that is updated automatically when this Dynamic is changed.

Each time a value is updated in this Dynamic, map will be invoked and the result will be stored in the returned Dynamic.

None is returned if this value belonged to a scope that has been destructed.

pub fn for_each<F>(&self, for_each: F)where F: for<'a> FnMut(&'a T) + Send + 'static,

Attaches for_each to this dynamic such that it is executed each time this dynamic changes.

pub fn generation(&self) -> Option<NonZeroUsize>

Returns the current generation of this dynamic.

A generation is a counter that is updated each time the value is accessed mutably.